requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
pytest-mock==3.10.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
